SEQNG-607 Split some classes on the client model
